fix(dialect): quote newlines in string literals across dialects
Two problems:
1) dialects didn't quote newlines properly
2) tests weren't complete or in the right places
